requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
amqp==5.3.1
anyscale==0.26.39
asgiref==3.8.1
attrs==25.3.0
awsebcli==3.25
bcrypt==4.3.0
billiard==4.2.1
blessed==1.21.0
boto3==1.39.3
botocore==1.39.3
cachetools==5.5.2
celery==5.5.3
cement==2.10.14
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==45.0.5
decorator==5.2.1
Deprecated==1.2.18
dj-database-url==1.3.0
Django==5.2.3
django-celery-beat==2.8.1
django-timezone-field==7.1
fabric==3.2.2
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-certificate-manager==1.10.2
google-cloud-common==1.5.2
google-cloud-compute==1.31.0
google-cloud-core==2.4.3
google-cloud-filestore==1.13.2
google-cloud-redis==2.18.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.73.1
gunicorn==23.0.0
httplib2==0.22.0
humanize==4.12.3
idna==3.10
invoke==2.2.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
log-symbols==0.0.14
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.6.3
numpy==2.2.6
oauth2client==4.1.3
opencv-python==4.12.0.88
packaging==24.2
pandas==2.3.0
paramiko==3.5.1
pathspec==0.12.1
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyNaCl==1.5.0
pyparsing==3.2.3
pytesseract==0.3.13
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.0
semantic-version==2.10.0
setuptools==80.9.0
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
spinners==0.0.24
sqlparse==0.5.3
tabulate==0.9.0
termcolor==2.5.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.9.0
wrapt==1.17.2
yarl==1.20.1
